Ground checks for jumping
There’s now a floor check done via trigger so that you can’t jump again
when in mid-air. Problems notices: It’s hard as fuck to get a good
running jump off a platform edge. This is due to the interpolation.
Turn it off and it’s easy, but then movement looks fucking awful. Also
problematic is the 3 blocks to the player’s left. They’re 3 separate
blocks of identical size at an identical height, yet somehow, one of
them occasionally collides with the player. This should not be
possible, but somehow it’s happening, and it’s fucking annoying.